[🚚rename]: common 폴더의 네이밍을 Shared로 변경
