+++ src/components/CodeArchive/CodeArchiveItem.tsx
@@ -3,8 +3,8 @@ import { css } from '@emotion/react';
 import { theme, ThemeType } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
 
-import Label from '../common/Label';
 import CodeBlock from '../gpt/CodeBlock';
+import Label from '../Shared/Label';
 
 type CodeArchiveItemProps = {
   requirement: RequirementStateType;
