+++ src/components/KanbanBoard/KanbanCard.tsx
@@ -2,7 +2,7 @@ import { css } from '@emotion/react';
 
 import { IssueStateType, IssueStatusType } from '@/types/issue';
 
-import Label from '../common/Label';
+import Label from '../Shared/Label';
 import KanbanDroppable from './KanbanDroppable';
 
 type KanbanCardProps = {
