+++ src/components/Requirement/RequirementCheckToggle.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from 'react';
 import { useRequirement } from '@/hooks/useRequirement';
 import { RequirementStateType } from '@/types/requirement';
 
-import CheckLottie from '../common/CheckLottie';
+import CheckLottie from '../Shared/CheckLottie';
 import CheckCircleIcon from '../icons/CheckCircleIcon';
 import EmptyCircleIcon from '../icons/EmptyCircleIcon';
