+++ src/components/Shared/CheckLottie.tsx
