+++ src/components/Shared/ResizeHandle.tsx
