+++ src/pages/_layout.tsx
@@ -1,10 +1,10 @@
 import { Outlet } from 'react-router-dom';
 
-import AsyncBoundary from '@/components/common/AsyncBoundary';
-import Header from '@/components/common/Header';
-import Navbar from '@/components/common/Navbar';
-import SkeletonBoard from '@/components/common/SkeletonBoard';
-import ToastContainer from '@/components/common/ToastContainer';
+import AsyncBoundary from '@/components/Shared/AsyncBoundary';
+import Header from '@/components/Shared/Header';
+import Navbar from '@/components/Shared/Navbar';
+import SkeletonBoard from '@/components/Shared/SkeletonBoard';
+import ToastContainer from '@/components/Shared/ToastContainer';
 import ModalContainer from '@/components/Modal/ModalContainer';
 import { useToggle } from '@/hooks/useToggle';
